+++ apps/heureka/src/types/index.d.ts
@@ -5,3 +5,13 @@
 
 /// <reference types="vite-plugin-svgr/client" />
 /// <reference types="vite/client" />
+
+declare module "@cloudoperators/juno-app-heureka" {
+  export type PluginModule = {
+    mount: (container: HTMLElement, options?: Record<string, any>) => void
+    unmount: () => void
+  }
+
+  export function mount(container: HTMLElement, options?: Record<string, any>): void
+  export function unmount(): void
+}