+++ packages/ui-components/src/components/Button/Button.component.tsx
@@ -1,9 +1,17 @@
 /*
- * SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and Juno contributors
+ * SPDX-FileCopyrightText: 2026 SAP SE or an SAP affiliate company and Juno contributors
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import React, { forwardRef, HTMLProps, LegacyRef, MouseEventHandler, ReactNode, MouseEvent } from "react"
+import React, {
+  forwardRef,
+  HTMLProps,
+  MouseEventHandler,
+  ReactNode,
+  MouseEvent,
+  ButtonHTMLAttributes,
+  Ref,
+} from "react"
 import { Icon } from "../Icon/index"
 import { KnownIcons } from "../Icon/Icon.component"
 import { Spinner } from "../Spinner/index"
@@ -38,24 +46,24 @@ const btnDefaultBase = `
 `
 
 const btnSmallDefaultPadding = `
-  jn:py-[0.3125rem]
+  jn:py-1.25
   jn:px-[0.5rem]
 `
 
 const btnSmallSubduedPadding = `
   jn:py-[0.25rem]
-  jn:px-[0.4375rem]
+  jn:px-1.75
 `
 
 // default size padding
 const btnDefaultPadding = `
-  jn:py-[0.4375rem]
-  jn:px-[0.625rem] 
+  jn:py-1.75
+  jn:px-2.5 
 `
 
 const btnDefaultSubduedPadding = `
-  jn:py-[0.375rem]
-  jn:px-[0.5625rem]
+  jn:py-1.5
+  jn:px-2.25
 `
 
 const defaultButtonColors = `
@@ -166,6 +174,7 @@ export const Button = forwardRef<HTMLButtonElement | HTMLAnchorElement, ButtonPr
     {
       label,
       title,
+      type = "button",
       variant = "default",
       size = "default",
       disabled,
@@ -183,7 +192,7 @@ export const Button = forwardRef<HTMLButtonElement | HTMLAnchorElement, ButtonPr
     const titleValue = title || label || ""
 
     const buttonIcon = progress ? (
-      <Spinner size={size === "small" ? "1.125rem" : "1.5rem"} color={`${spinnerColorClass(variant)}`} />
+      <Spinner size={size === "small" ? "1.125rem" : "1.5rem"} color={spinnerColorClass(variant)} />
     ) : icon ? (
       <Icon
         icon={icon}
@@ -213,9 +222,9 @@ export const Button = forwardRef<HTMLButtonElement | HTMLAnchorElement, ButtonPr
           ${className}`}
         disabled={disabled || undefined}
         onClick={handleClick}
-        ref={ref as LegacyRef<HTMLButtonElement>}
+        ref={ref as Ref<HTMLButtonElement>}
         {...(props as HTMLProps<HTMLButtonElement>)}
-        type="button"
+        type={type}
         title={titleValue}
       >
         {buttonIcon}
@@ -240,7 +249,7 @@ export const Button = forwardRef<HTMLButtonElement | HTMLAnchorElement, ButtonPr
         `}
         disabled={disabled}
         onClick={onClick}
-        ref={ref as LegacyRef<HTMLAnchorElement>}
+        ref={ref as Ref<HTMLAnchorElement>}
         {...(props as HTMLProps<HTMLAnchorElement>)}
         title={titleValue}
       >
@@ -311,6 +320,12 @@ export interface ButtonProps extends Omit<HTMLProps<HTMLAnchorElement> | HTMLPro
    * Click event handler for the button.
    */
   onClick?: MouseEventHandler<HTMLButtonElement | HTMLAnchorElement>
+
+  /**
+   * Button type.
+   * @default "button"
+   */
+  type?: ButtonHTMLAttributes<HTMLButtonElement>["type"]
 
   /**
    * Indicates whether the button's action is in progress.